Add CorePython, CoreSync and sync command
Introduce Python implementations for core and sync protocols and add a CLI sync entrypoint. Creates swanlab/sdk/internal/core_python/core.py (CorePython implementation) and swanlab/sdk/internal/core_python/sync.py (CoreSyncPython stub), plus a new swanlab/sdk/cmd/sync.py command. Update __init__.py to export CorePython and CoreSyncPython, and extend swanlab/sdk/protocol/core.py to import the sync proto and declare CoreSyncProtocol. These changes scaffold sync-related functionality and centralize protocol exports; methods largely stubbed pending further implementation.
