+++ swanlab/sdk/internal/pkg/helper/__init__.py
@@ -8,6 +8,7 @@
 from pathlib import Path, PurePosixPath
 
 from .env import DEBUG, is_interactive, is_jupyter
+from .impl import get_core_impl, get_probe_impl
 from .system import fmt_system_key, is_system_key
 from .version import get_swanlab_latest_version, get_swanlab_version
 
@@ -21,6 +22,8 @@ __all__ = [
     "is_system_key",
     "get_swanlab_version",
     "get_swanlab_latest_version",
+    "get_probe_impl",
+    "get_core_impl",
     "mkdir_and_append_gitignore",
 ]
